src/requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-cors-headers==2.1.0
django-rest-auth==0.9.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
